+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 CHANGELOG
 =========
 
+3.0.2
+- Added file loader to `ConfigFactory` for those times when you want to load the array first before creating the `Config` object.
+- Refactored tests that used non PHPUnit 4.8 asserts.
+- Added TravisCI.
+- Changed to MIT license.
+
 3.0.0
 - initial release